+++ packages/svelte/src/internal/index.js
@@ -36,7 +36,9 @@ export {
 	pop,
 	push,
 	reactive_import,
-	effect_active
+	effect_active,
+	user_root_effect,
+	inspect
 } from './client/runtime.js';
 
 export * from './client/validate.js';